[PTC] Move ControlScreen in a separate Class

+++ src/Player/HUD.cpp
@@ -9,18 +9,15 @@
 
 HUD::HUD()
 {
-    for (int i = 1; i < 5; i++)
+    for (size_t i = 1; i < 5; i ++)
         battery.push_back(std::make_shared<Battery>(i, 3));
-    for (int i = 0; i < 3; i ++) {
+    for (size_t i = 0; i < 3; i ++) {
         heart.push_back(std::make_shared<ImageSFML>("resources/Images/Hud/heart.png"));
         heart[i]->setScale(sf::Vector2f(0.25, 0.25));
     }
     weapon_type = 1;
 }
-
-HUD::~HUD()
-{
-}
+HUD::~HUD() {}
 
 void HUD::display(std::shared_ptr<sf::RenderWindow> window, int weapon_type, int life) {
     int x = 800;
@@ -57,11 +54,11 @@ void HUD::display(std::shared_ptr<sf::RenderWindow> window, int weapon_type, int
         battery[weapon_type]->display(window, {window->getView().getCenter().x - 600, window->getView().getCenter().y + 450});
     else
         battery[0]->display(window, {window->getView().getCenter().x - 600, window->getView().getCenter().y + 450});
-    for (int i = 0; i < life; i++) {
+    for (size_t i = 0; i < life; i ++) {
         heart[i]->setPosition(sf::Vector2f(window->getView().getCenter().x - x,
             window->getView().getCenter().y + 310));
         window->draw(heart[i]->getSprite());
-        x+= 50;
+        x += 50;
     }
 }
 
@@ -83,27 +80,12 @@ void HUD::displaySprintBar(std::shared_ptr<sf::RenderWindow> window, int sprintB
 
 void HUD::incWeaponType()
 {
-    weapon_type++;
+    weapon_type ++;
     if (weapon_type > 4)
         weapon_type = 1;
 }
 
-void HUD::incMun(int batteryNb)
-{
-    battery[batteryNb]->incMaxMun();
-}
-
-int HUD::getMunBattery()
-{
-    return (battery[weapon_type -1]->getMun());
-}
-
-int HUD::decBatteryMun()
-{
-    return (battery[weapon_type - 1]->decMun());
-}
-
-void HUD::batteryChanneling()
-{
-    battery[weapon_type - 1]->channeling();
-}
+void HUD::incMun(int batteryNb) {battery[batteryNb]->incMaxMun();}
+int HUD::getMunBattery() {return (battery[weapon_type -1]->getMun());}
+int HUD::decBatteryMun() {return (battery[weapon_type - 1]->decMun());}
+void HUD::batteryChanneling() {battery[weapon_type - 1]->channeling();}
